pyproject.toml
[tool.poetry]
name = "pywechat"
version = "0.1.0"
description = "Python client for WeChat"
authors = ["Jiang Yue <maze1024@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.2"
pydantic = "^2.9.2"
cryptography = "^43.0.1"
xmltodict = "^0.13.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
fastapi = "^0.115.0"
python-dotenv = "^1.0.1"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-multipart = "^0.0.10"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
